Add files via upload
This is the main script that runs the survey expander, it is equipped with functions that can get the mean and std of columns, or the average values of certain columns. Or even columns where the users have to rank 1st -5th. It is able to replicate previous results and mimic them.